Arcyd: workaround bad unicode in diffs
Sometimes source files that are being diffed will contain bad unicode
sequences; in these cases we still want to be able to create a review.

As a first step we'll simply replace the bad unicode characters inline
in the diff.  Later we'll want to add a warning on the review page that
we've done this.

Test Plan: $ ./precommit.sh
(no coverage of bad unicode path)