Arcyd: implement the r-branch proposal
Implement a new branch naming scheme to enable the creation of branches
named like this:

  r/base/description

Which is shorter and easier to type than the classic scheme and also
allows branches with the same base to sort together.

Add a trivial test driver.

Add a repetition in the acceptance tests with the new naming scheme.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh