Puppet: download older code, as zips from GitHub
GitHub is now incompatible with Git versions less than 1.7.10 when
cloning via https, workaround this for machines with older Git clients
by downloading the code as zips instead.

Recent versions of Phabricator won't work out of the box with our
initial.db and require additional setup of auth methods.  Work around
this by targetting an older version which is known to be compatible.

Split the creation of the /phabricator/instances/dev dir off into a
separate class and make 'phabricator' depend on it.  Apparently it was
working as expected by happy coincidence before.

Add an 'https_proxy' var as this is now necessary for getting the code.

Fixes GitHub issue bloomberg/phabricator-tools#4

Test Plan:
On Windows / msysGit, with proxy settings set
On OSX10, without proxy settings
    phabricator-tools/ $ cd vagrant
    vagrant/ $ vagrant destroy
    vagrant/ $ vagrant up
    browse to 'http://127.0.0.1:8080' and login as phab/password