phlgitu_ref: add Name, is_remote(), is_fq()
Continue making it easier to work with git remote names by providing a
vocabulary type for passing them around unambiguously with support for
conversion to convenient forms at point of usage.

This should reduce the amount of care needed when passing git references
around, it's only important to concentrate when the reference name is
actually consumed.

Test Plan:
$ ./precommit.sh