Add phlsys_fs.lockfile_context()
Support exclusive operations on the filesystem in a similar way to git -
using lockfiles.

The drawback to this is that it's possible to 'leak' the lockfile if the
program is forcibly terminated or exits at 'just the wrong time'.

Test Plan:
exercise lockfile_context() using testbed/lockfile/run.sh, this
simulates multiple writers contending for the same resource in a tight
loop.

Observed that occasionally a lockfile is left behind after killing the
writer processes.