arcyd/acceptance...: test pruning landing log
Add a test to the arcyd acceptance tests to make sure that we can
manually prune the landing log and have that change persist across new
reviews.

Previously this may have gone wrong because arcyd wasn't fetching the
special refs with a refspec beginning with '+', which enables non-ff
updates.

Test Plan:
$ ./testbed/arcyd/acceptance_tests.sh