Arcyd: refactor, mv repo config parsing to 'start'
Make it possible to do all disk access to repository config files for
'arcyd start' before doing abdi_processrepos.process().

This makes it possible to get a lock on the repository config files
at the top-level of the program and release it before descending into
the arcyd daemon details.

Test Plan:
$ ./precommit.sh