abdt_conduit uses cached author details
Cache expensive lookups on the author user, using the new cached
author_phid in ReviewStateCache and UsernamePhidCache.get_username().

Write the new author through to the cache when commandeering a review.

Exercise the author user cache by abandoning the review after
commandeering in the breathing tests.

Test Plan:

Exercise abdt_conduit as much as we can with
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh

Deliberately break the caching temporarily to make sure the breathing
test fails as expected.