Add phlcon_user.UsernamePhidCache.get_username()
Make it possible to also query username->phid with the
UsernamePhidCache, this will be useful in Arcyon and Arcyd; both to
introduce new caching and to simplify existing caching.

Update query_users_from_phids() to raise OneOrMoreUnknownPhids so that
the condition can be handled explicitly.

Add test code to cover new UsernamePhidCache functionality.

Test Plan:

Update doc, exercise changed code with
$ ./precommit.sh

Exercise a bit more to be sure
$ ./testbed/arcyd/acceptance_tests.sh