abdi_processrepoarglist: separate review cache
Store the review cache alongside the arcyd conduit, so that we may
operate on it directly.

This will allow us to later merge the caches from worker processes back
together at the end of each cycle, so we know what needs to be batch
queried at the start of each cycle.

Test Plan:
$ ./precommit.sh