phlcon_reviewstatecache: refactor, rm duplication
Replace the useless proxy class 'ReviewStateCache' with a factory
function. The only value it adds is in it's constructor, the rest is
pure duplication overhead.

We'll rename the old imp detail _ReviewStateCache to ReviewStateCache
proper in a subsequent commit.

Test Plan:
Make sure we got all the replacements
$ git grep '\bReviewStateCache\b'

General coverage
$ ./precommit.sh