phlcon_reviewstatecache: expose active review set
Make it possible to retrieve and merge the active review set, this will
allow the Arcyd client to merge the active review sets after running in
multiple worker processes.

This will also allow the Arcyd client to hint the revisions that
should be loaded for each repository before the first run through the
update loop.

Add tests to cover this.

Test Plan:
$ ./precommit.sh