abdi_processrepoarglist: sleep to pad out cycle
Instead of unconditionally sleeping for a set number of seconds at the
end of each cycle, sleep only if the cycle time is less than the
specified time to sleep.

This means that if the cycles are too long, they don't get made longer
by sleeping in addition.

Test Plan:
$ ./precommit.sh

.. log the seconds to sleep ..
$ ./testbed/arcyd/test_shell.sh
    .. monitor the sleep secs under different conditons ..