Arcyd: refactor, inject refcache_repo to abdt_git
Make it possible to interact with the underlying phlgitx_refcache repo
separately from a abdt_git.Repo.

This is useful because we'll be able to manipulate the underlying
refcache without needing to widen the contract of the abdt_git.Repo.

We'll be able to use this to ensure that cache updates from worker
processes are propagated back to the original caches in the main
process.

Test Plan:
general coverage
$ ./precommit.sh