Arcyd: refactor, re-use abdi_repoargs phab parser
Remove duplication of argument parsing by re-using
abdi_repoargs.setup_repohost_parser in abdcmd_addphabricator

Note that this makes the '--review-url-format' argument optional, but
also the value of this argument is not used anywhere and will be removed
in a later commit. This later commit will be marked as a breaking
change.

Test Plan:
$ ./precommit.sh