Arcyd: fix caching for --no-loop and max-workers=1
It turns out that we weren't consuming the 'newness' of the snoop urls
in the case where we're not multiprocessing and in the case where we're
not looping.

In the non-multiprocessing case, we simply weren't consuming the
newness.

In the --no-loop case, we weren't writing the cache to disk.

This resulted in fetching more times than was necessary.

Test Plan:
$ ./precommit

$ ./testbed/arcyd/test_shell.sh
    $ sed -i '/--max-workers/{n;s/0/1/}' ../arcyd/configfile
    $ grep -A 1 -- '--max-workers' ../arcyd/configfile
    --max-workers
    1
    $ cd ../arcyd
    $ arcyd stop
    $ ~/repos/phabricator-tools/proto/arcyd stop
    $ cat .arcyd.urlwatcher.cache
    ... ["8296dc818c89de923afab693bfefc2e2a26f3081", false]}
    $ cd ../dev
    $ ./poke.sh
    $ cd ../arcyd
    $ ~/repos/phabricator-tools/proto/arcyd start --no-loop
    $ cat .arcyd.urlwatcher.cache
    ... ["e99c3848d7ed9b438896e478cd39de21cacc22b6", false]}