Add docker/gitdaemon for easy git experiments
In order to locally test docker containers that use remote git
repositories, it will be very useful to be able to host a simple git
server in a docker container.

Add a Dockerfile and entrypoint that will implement a simple git server
with git-daemon.

Note that the list of repositories that it hosts is currently fixed at
the time of starting the container.

Test Plan:

$ cd docker
$ docker build -t gitdaemon .
$ docker run -d --name mygit gitdaemon a b c

$ docker run -it --rm alpine sh
    # apk update
    # apk add git
    # git clone git://mygit/a
    # cd a
    # touch README
    # git add README
    # git config user.email me
    # git config user.name me
    # git commit -m 'Initial commit'
    # git push