Add docker/arcyd/init to reap zombies
The process which is pid1 in a container should wait() on any zombies
that the kernel re-parents to it, otherwise they will accumulate.

Create a new executable to take care of these zombies, which is also
responsible for passing arguments to 'entrypoint', passing on SIGTERM
and waiting for everything to exit.

Test Plan:
$ ./precommit.sh

$ ./testbed/arcyd/docker-testshell.sh
    $ docker stop arcydtest-arcyd
    $ docker logs arcydtest-arcyd
    .. Note that it stopped as expected ..

.. Introduce a leak of subprocesses in entrypoint ..
$ ./testbed/arcyd/docker-testshell.sh
    $ docker stop arcydtest-arcyd
    $ docker logs arcydtest-arcyd
    .. Note that it stopped as expected ..

$ cd docker/arcyd
    $ ../build-image.sh arcyd-dockerfile arcyd
    $ docker run --rm -it arcyd
    $ echo $?
    2